+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/NSQLQuery.java
@@ -1,5 +1,10 @@
 package com.dwarfeng.fdr.impl.dao;
 
+import com.dwarfeng.subgrade.stack.exception.DaoException;
+import org.springframework.lang.NonNull;
+
+import java.sql.Connection;
+
 /**
  * 本地SQL查询。
  *
@@ -9,4 +14,9 @@ package com.dwarfeng.fdr.impl.dao;
 public interface NSQLQuery {
 
     boolean supportType(String type);
+
+    /**
+     * @since 1.9.4
+     */
+    void init(@NonNull Connection connection) throws DaoException;
 }