+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/PersistenceValueNSQLQuery.java
@@ -20,7 +20,8 @@ public interface PersistenceValueNSQLQuery extends NSQLQuery {
 
     List<PersistenceValue> lookupPersistence(@NonNull Connection connection, Object[] objs) throws DaoException;
 
-    List<PersistenceValue> lookupPersistence(@NonNull Connection connection, Object[] objs, PagingInfo pagingInfo) throws DaoException;
+    List<PersistenceValue> lookupPersistence(@NonNull Connection connection, Object[] objs, PagingInfo pagingInfo)
+            throws DaoException;
 
     Integer lookupPersistenceCount(@NonNull Connection connection, Object[] objs) throws DaoException;