+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/nsql/MySQL8PersistenceValueNSQLQuery.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.fdr.impl.dao.nsql;
 
 import com.dwarfeng.fdr.impl.dao.PersistenceValueNSQLQuery;
+import com.dwarfeng.fdr.sdk.util.Constants;
 import com.dwarfeng.fdr.stack.bean.entity.PersistenceValue;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
@@ -28,14 +29,10 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
     }
 
     @Override
-    public List<PersistenceValue> lookupPersistence(
-            @NonNull Connection connection, Object[] objs
-    ) throws DaoException {
+    public List<PersistenceValue> lookupPersistence(@NonNull Connection connection, Object[] objs) throws DaoException {
         try {
             NSQLQueryUtil.objsValidation(
-                    objs,
-                    new Class[]{LongIdKey.class, Date.class, Date.class},
-                    new boolean[]{true, false, false}
+                    objs, new Class[]{LongIdKey.class, Date.class, Date.class}, new boolean[]{true, false, false}
             );
             LongIdKey pointKey = (LongIdKey) objs[0];
             Date startDate = (Date) objs[1];
@@ -43,6 +40,7 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
 
             StringBuilder sqlBuilder = new StringBuilder();
             selectTable(sqlBuilder);
+            MySQL8NSQLQueryUtil.forceIndex(sqlBuilder, Constants.INDEX_NAME_POINT_ID_HAPPENED_DATE);
             sqlBuilder.append("WHERE ");
             {
                 if (Objects.isNull(pointKey)) {
@@ -90,9 +88,7 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
     ) throws DaoException {
         try {
             NSQLQueryUtil.objsValidation(
-                    objs,
-                    new Class[]{LongIdKey.class, Date.class, Date.class},
-                    new boolean[]{true, false, false}
+                    objs, new Class[]{LongIdKey.class, Date.class, Date.class}, new boolean[]{true, false, false}
             );
             LongIdKey pointKey = (LongIdKey) objs[0];
             Date startDate = (Date) objs[1];
@@ -100,6 +96,7 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
 
             StringBuilder sqlBuilder = new StringBuilder();
             selectTable(sqlBuilder);
+            MySQL8NSQLQueryUtil.forceIndex(sqlBuilder, Constants.INDEX_NAME_POINT_ID_HAPPENED_DATE);
             sqlBuilder.append("WHERE ");
             {
                 if (Objects.isNull(pointKey)) {
@@ -147,14 +144,10 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
     }
 
     @Override
-    public Integer lookupPersistenceCount(
-            @NonNull Connection connection, Object[] objs
-    ) throws DaoException {
+    public Integer lookupPersistenceCount(@NonNull Connection connection, Object[] objs) throws DaoException {
         try {
             NSQLQueryUtil.objsValidation(
-                    objs,
-                    new Class[]{LongIdKey.class, Date.class, Date.class},
-                    new boolean[]{true, false, false}
+                    objs, new Class[]{LongIdKey.class, Date.class, Date.class}, new boolean[]{true, false, false}
             );
             LongIdKey pointKey = (LongIdKey) objs[0];
             Date startDate = (Date) objs[1];
@@ -169,6 +162,7 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
             {
                 sqlBuilder.append("tbl_persistence_value AS tbl ");
             }
+            MySQL8NSQLQueryUtil.forceIndex(sqlBuilder, Constants.INDEX_NAME_POINT_ID);
             sqlBuilder.append("WHERE ");
             {
                 if (Objects.isNull(pointKey)) {
@@ -199,12 +193,12 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
     }
 
     @Override
-    public PersistenceValue previous(
-            @NonNull Connection connection, LongIdKey pointKey, Date date
-    ) throws DaoException {
+    public PersistenceValue previous(@NonNull Connection connection, LongIdKey pointKey, Date date)
+            throws DaoException {
         try {
             StringBuilder sqlBuilder = new StringBuilder();
             selectTable(sqlBuilder);
+            MySQL8NSQLQueryUtil.forceIndex(sqlBuilder, Constants.INDEX_NAME_POINT_ID_HAPPENED_DATE_DESC);
             sqlBuilder.append("WHERE ");
             {
                 if (Objects.isNull(pointKey)) {
@@ -245,12 +239,11 @@ public class MySQL8PersistenceValueNSQLQuery extends AbstractNSQLQuery implement
     }
 
     @Override
-    public PersistenceValue rear(
-            @NonNull Connection connection, LongIdKey pointKey, Date date
-    ) throws DaoException {
+    public PersistenceValue rear(@NonNull Connection connection, LongIdKey pointKey, Date date) throws DaoException {
         try {
             StringBuilder sqlBuilder = new StringBuilder();
             selectTable(sqlBuilder);
+            MySQL8NSQLQueryUtil.forceIndex(sqlBuilder, Constants.INDEX_NAME_POINT_ID_HAPPENED_DATE);
             sqlBuilder.append("WHERE ");
             {
                 if (Objects.isNull(pointKey)) {